refactor: adjust code position
