+++ src/components/projects/policy/home.vue
@@ -14,13 +14,20 @@
         </span>
       </el-tab-pane>
     </el-tabs>
-    <Policy></Policy>
+    <Policy :userList="userList" :workflowList="workflowList" :envList="envList" :policy="policy"></Policy>
   </div>
 </template>
 
 <script>
 import Policy from './policy.vue'
 import bus from '@utils/event_bus'
+import {
+  usersAPI,
+  queryPolicyDefinitionsAPI,
+  getWorkflowsInProjectAPI,
+  getCommonWorkflowListAPI,
+  listProductAPI
+} from '@api'
 export default {
   data () {
     return {
@@ -29,7 +36,14 @@ export default {
           name: '',
           initName: ''
         }
-      ]
+      ],
+      userList: [],
+      workflowList: [],
+      envList: [],
+      policy: {
+        workflow: [],
+        environment: []
+      }
     }
   },
   computed: {
@@ -51,12 +65,84 @@ export default {
         name: '',
         initName: `zadig-add-mode-${this.allMode.lastLength}`
       })
+    },
+    getUsers () {
+      const projectName = this.projectName
+      const payload = {
+        name: '',
+        page: 1,
+        per_page: 1000000
+      }
+      usersAPI(payload, projectName).then(res => {
+        this.userList = _.uniqBy(res.users, 'uid')
+      })
+    },
+    async getPolicyDefinitions () {
+      const res = await queryPolicyDefinitionsAPI(
+        this.projectName
+      ).catch(error => console.log(error))
+      if (res) {
+        Object.keys(this.policy).forEach(key => {
+          this.policy[key] = res
+            .find(group => group.resource.toLowerCase() === key)
+            .rules.filter(rule => !rule.action.startsWith('create_'))
+            .map(rule => {
+              return {
+                ...rule,
+                icon: this.selectIcon(rule.action)
+              }
+            })
+        })
+      }
+    },
+    selectIcon (rule) {
+      const iconEnum = {
+        get_: 'el-icon-view',
+        edit_: 'el-icon-edit-outline',
+        delete_: 'el-icon-delete',
+        run_: 'el-icon-video-play',
+        config_: 'el-icon-setting',
+        manage_: 'el-icon-menu'
+      }
+      const icon = Object.keys(iconEnum).find(key => rule.startsWith(key))
+      return icon ? iconEnum[icon] : ''
+    },
+    async getWorkflows () {
+      let res = []
+      res = await getWorkflowsInProjectAPI(this.projectName).catch(err => {
+        console.log(err)
+        return []
+      })
+      const workflowList = await getCommonWorkflowListAPI(
+        this.projectName
+      ).catch(err => {
+        console.log(err)
+        return []
+      })
+      workflowList.workflow_list.forEach(list => {
+        list.type = 'common'
+      })
+      this.workflowList = [...res, ...workflowList.workflow_list]
+        .filter(workflow => !workflow.base_name)
+        .map(workflow => workflow.name)
+    },
+    async getEnvNameList () {
+      const res = await listProductAPI(this.projectName).catch(err =>
+        console.log(err)
+      )
+      if (res) {
+        this.envList = res.filter(env => !env.base_name).map(env => env.name)
+      }
     }
   },
   components: {
     Policy
   },
   created () {
+    this.getUsers()
+    this.getPolicyDefinitions()
+    this.getWorkflows()
+    this.getEnvNameList()
     bus.$emit(`set-topbar-title`, {
       title: '',
       breadcrumb: [
